If there are mzmine files from another user into the temporal folder and you don't have right to write them, the program stops. A new condition is added to remove the files only if they can be written by the user.

+++ src/net/sf/mzmine/main/mzmineclient/MZmineClient.java
@@ -16,7 +16,6 @@
  * MZmine 2; if not, write to the Free Software Foundation, Inc., 51 Franklin St,
  * Fifth Floor, Boston, MA 02110-1301 USA
  */
-
 package net.sf.mzmine.main.mzmineclient;
 
 import java.io.File;
@@ -45,7 +44,6 @@ import org.dom4j.io.SAXReader;
 public class MZmineClient extends MZmineCore implements Runnable {
 
 	private Logger logger = Logger.getLogger(this.getClass().getName());
-
 	private Vector<MZmineModule> moduleSet;
 
 	// make MZmineClient a singleton
@@ -86,27 +84,30 @@ public class MZmineClient extends MZmineCore implements Runnable {
 
 			// Find all files with the mask mzmine*.scans
 			File remainingTmpFiles[] = tempDir.listFiles(new FilenameFilter() {
+
 				public boolean accept(File dir, String name) {
 					return name.matches("mzmine.*\\.scans");
 				}
 			});
 
 			for (File remainingTmpFile : remainingTmpFiles) {
+				if (remainingTmpFile.canWrite()) {
+					// Try to obtain a lock on the file
+					RandomAccessFile rac = new RandomAccessFile(remainingTmpFile,
+							"rw");
 
-				// Try to obtain a lock on the file
-				RandomAccessFile rac = new RandomAccessFile(remainingTmpFile,
-						"rw");
-				FileLock lock = rac.getChannel().tryLock();
-				rac.close();
+					FileLock lock = rac.getChannel().tryLock();
+					rac.close();
 
-				if (lock != null) {
-					// We locked the file, which means nobody is using it
-					// anymore and it can be removed
-					logger.finest("Removing unused file " + remainingTmpFile);
-					remainingTmpFile.delete();
+					if (lock != null) {
+						// We locked the file, which means nobody is using it
+						// anymore and it can be removed
+						logger.finest("Removing unused file " + remainingTmpFile);
+						remainingTmpFile.delete();
+					}
 				}
-
 			}
+
 
 			SAXReader reader = new SAXReader();
 			configuration = reader.read(CONFIG_FILE);
@@ -118,16 +119,15 @@ public class MZmineClient extends MZmineCore implements Runnable {
 			String numberOfNodesConfigEntry = null;
 			Element nodesElement = configRoot.element(NODES_ELEMENT_NAME);
 			if (nodesElement != null) {
-				numberOfNodesConfigEntry = nodesElement
-						.attributeValue(LOCAL_ATTRIBUTE_NAME);
+				numberOfNodesConfigEntry = nodesElement.attributeValue(LOCAL_ATTRIBUTE_NAME);
 			}
 			if (numberOfNodesConfigEntry != null) {
 				numberOfNodes = Integer.parseInt(numberOfNodesConfigEntry);
-			} else
+			} else {
 				numberOfNodes = Runtime.getRuntime().availableProcessors();
+			}
 
-			logger.info("MZmine starting with " + numberOfNodes
-					+ " computation nodes");
+			logger.info("MZmine starting with " + numberOfNodes + " computation nodes");
 
 			logger.finer("Loading core classes");
 
@@ -152,19 +152,16 @@ public class MZmineClient extends MZmineCore implements Runnable {
 
 			moduleSet = new Vector<MZmineModule>();
 
-			Iterator<Element> modIter = configRoot
-					.element(MODULES_ELEMENT_NAME).elementIterator(
-							MODULE_ELEMENT_NAME);
+			Iterator<Element> modIter = configRoot.element(MODULES_ELEMENT_NAME).elementIterator(
+					MODULE_ELEMENT_NAME);
 
 			while (modIter.hasNext()) {
 				Element moduleElement = modIter.next();
-				String className = moduleElement
-						.attributeValue(CLASS_ATTRIBUTE_NAME);
+				String className = moduleElement.attributeValue(CLASS_ATTRIBUTE_NAME);
 				loadModule(className);
 			}
 
-			MZmineCore.initializedModules = moduleSet
-					.toArray(new MZmineModule[0]);
+			MZmineCore.initializedModules = moduleSet.toArray(new MZmineModule[0]);
 
 			MZmineCore.isLightViewer = false;
 
@@ -172,8 +169,7 @@ public class MZmineClient extends MZmineCore implements Runnable {
 			loadConfiguration(CONFIG_FILE);
 
 		} catch (Exception e) {
-			logger.log(Level.SEVERE, "Could not parse configuration file "
-					+ CONFIG_FILE, e);
+			logger.log(Level.SEVERE, "Could not parse configuration file " + CONFIG_FILE, e);
 			System.exit(1);
 		}
 
@@ -200,8 +196,7 @@ public class MZmineClient extends MZmineCore implements Runnable {
 			Class moduleClass = Class.forName(moduleClassName);
 
 			// create instance
-			MZmineModule moduleInstance = (MZmineModule) moduleClass
-					.newInstance();
+			MZmineModule moduleInstance = (MZmineModule) moduleClass.newInstance();
 
 			// init module
 			moduleInstance.initModule();
